Prof feedback revisions
You need to account for missing images. Search for movies with "array" in the title.   ---- This was an error of using the filepath relative to JS file instead of HTML

Usability - don't make the hover state of a link bold. That makes the text wider on hover and things can shift around on the page due to the width difference. ---- I was on the fence about this anyways. It didn't look great

If a bad search is performed (search for "sfgbsegsbeb") - it should not be saved, there's should not be a link to "don't remember...", and the previous good search should be deleted from storage. --- Simple fix of adding a localstorage clear in the check for results or not, as well as remove() for the rememberSpan